+++ mods/Farming/mod/prototypes/recipe/recipe-updates.lua
@@ -4,11 +4,11 @@ require("__RsUtility__.scripts.prototypes")
 local patch_recipe = data.raw.recipe[get_vanilla_name("rs-farming-patch")]
 for index, ingredient in pairs(patch_recipe.ingredients) do
   if ingredient[1] == get_vanilla_name("rs-plank") then
-    if not data.raw.item[ingredient[1]] then
+    if not data.raw.item[ingredient[1]] and not data.raw.tool[ingredient[1]] then
       ingredient[1] = "wood"
     end
   elseif ingredient[1] == get_vanilla_name("rs-bronze-nails") then
-    if not data.raw.item[ingredient[1]] then
+    if not data.raw.item[ingredient[1]] and not data.raw.tool[ingredient[1]] then
       patch_recipe.ingredients[index] = nil
     end
   end
@@ -23,11 +23,11 @@ end
 local bin_recipe = data.raw.recipe[get_vanilla_name("rs-compost-bin")]
 for index, ingredient in pairs(bin_recipe.ingredients) do
   if ingredient[1] == get_vanilla_name("rs-plank") then
-    if not data.raw.item[ingredient[1]] then
+    if not data.raw.item[ingredient[1]] and not data.raw.tool[ingredient[1]] then
       ingredient[1] = "wood"
     end
   elseif ingredient[1] == get_vanilla_name("rs-bronze-nails") then
-    if not data.raw.item[ingredient[1]] then
+    if not data.raw.item[ingredient[1]] and not data.raw.tool[ingredient[1]] then
       bin_recipe.ingredients[index] = nil
     end
   end
@@ -37,11 +37,11 @@ end
 local bucket_recipe = data.raw.recipe[get_vanilla_name("rs-bucket")]
 for index, ingredient in pairs(bucket_recipe.ingredients) do
   if ingredient[1] == get_vanilla_name("rs-plank") then
-    if not data.raw.item[ingredient[1]] then
+    if not data.raw.item[ingredient[1]] and not data.raw.tool[ingredient[1]] then
       ingredient[1] = "wood"
     end
   elseif ingredient[1] == get_vanilla_name("rs-bronze-nails") then
-    if not data.raw.item[ingredient[1]] then
+    if not data.raw.item[ingredient[1]] and not data.raw.tool[ingredient[1]] then
       bucket_recipe.ingredients[index] = nil
     end
   end